+++ src/singletons/Settings.hpp
@@ -378,6 +378,14 @@ public:
         "/misc/twitch/messageHistoryLimit",
         800,
     };
+    IntSetting twitchMessageLogLimit = {
+        "/misc/twitch/messageLogLimit",
+        10000,
+    };
+    IntSetting twitchMessageOnScreenLimit = {
+        "/misc/twitch/messageOnScreenLimit",
+        1000,
+    };
 
     IntSetting emotesTooltipPreview = {"/misc/emotesTooltipPreview", 1};
     BoolSetting openLinksIncognito = {"/misc/openLinksIncognito", 0};